+++ src/ACME/PreCure/Textbook/DeliciousParty/Profiles.hs
@@ -12,6 +12,7 @@ girls =
   [ mkGirl "Yui Nagomi" "和実 ゆい"
   , mkGirl "Kokone Fuwa"  "芙羽 ここね"
   , mkGirl "Ran Hanamichi"  "華満 らん"
+  , mkGirl "Amane Kasai"  "菓彩 あまね"
   ]
 
 transformees :: [Transformee]
@@ -34,6 +35,37 @@ transformees =
       "キュアヤムヤム"
       ""
       introducesHerselfAs_CureYumYum
+  , mkTransformee
+      "Cure Finale"
+      ""
+      "キュアフィナーレ"
+      ""
+      introducesHerselfAs_CureFinale
+
+  , mkTransformee
+      "Cure Precious"
+      "Party Up Style"
+      "キュアプレシャス"
+      "パーティアップスタイル"
+      introducesHerselfAs_CurePrecious
+  , mkTransformee
+      "Cure Spicy"
+      "Party Up Style"
+      "キュアスパイシー"
+      "パーティアップスタイル"
+      introducesHerselfAs_CureSpicy
+  , mkTransformee
+      "Cure Yum-Yum"
+      "Party Up Style"
+      "キュアヤムヤム"
+      "パーティアップスタイル"
+      introducesHerselfAs_CureYumYum
+  , mkTransformee
+      "Cure Finale"
+      "Party Up Style"
+      "キュアフィナーレ"
+      "パーティアップスタイル"
+      introducesHerselfAs_CureFinale
   ]
 
 transformedGroups :: [TransformedGroup]
@@ -53,7 +85,13 @@ specialItems =
   , mkSpecialItem "Pam-Pam"  "パムパム"
   , mkSpecialItem "Mem-Mem" "メンメン"
 
+  , mkSpecialItem "Heart Fruits Pendant" "ハートフルーツペンダント"
+  , mkSpecialItem "Creamy Fleuret" "クリーミーフルーレ"
+
   , mkSpecialItem "Heart Juicy Mixer"  "ハートジューシーミキサー"
+
+  , mkSpecialItem "Party Up Ring"  "パーティーアップリング"
+  , mkSpecialItem "Party Candle Tact"  "パーティーキャンドルタクト"
   ]
 
 
@@ -86,6 +124,16 @@ transformations =
       ++ kuruKuruMemMemMiracleMemMem
       ++ [sharin'Energy, wonton, memMem, introducesHerselfAs_CureYumYum]
       )
+  , mkTransformation
+      ["Amane"]
+      ["HeartFruitsPendant"]
+      ["CureFinale"]
+      ( preCureDeliciouStan'ByPartyGo
+      ++ fruitsFabulousOrder
+      ++ [sharin'Energy]
+      ++ toppingBrilliantShineMore
+      ++ [introducesHerselfAs_CureFinale]
+      )
 
   , mkTransformation
       ["Yui", "Kokone"]
@@ -114,12 +162,46 @@ transformations =
       ++ [introducesHerselfAs_CurePrecious, introducesHerselfAs_CureSpicy, introducesHerselfAs_CureYumYum]
       ++ ["デリシャスパーティ♡プリキュア！"]
       )
+
+  , mkTransformation
+      girls2
+      items2
+      groupMembers2
+      ( preCureDeliciouStan'ByPartyGo
+      ++ nigiNigiKomeKomeHeartWoKomeKome
+      ++ openPamPamSandPamPam
+      ++ kuruKuruMemMemMiracleMemMem
+      ++ fruitsFabulousOrder
+      ++ [sharin'Energy]
+      ++ [kome_, tasty, wonton]
+      ++ toppingBrilliantShineMore
+      ++ [komeKome, pamPam, memMem]
+      ++ [introducesHerselfAs_CurePrecious, introducesHerselfAs_CureSpicy, introducesHerselfAs_CureYumYum, introducesHerselfAs_CureFinale]
+      ++ ["デリシャスパーティ♡プリキュア！"]
+      )
+
+  , mkTransformation
+      groupMembers2
+      itemsLightMyDelicious
+      groupMembers_PartyUp
+      [ "（コメコメの力をみんなに！）"
+      , "パーティキャンドルタクト！"
+      , "笑顔のパワー！"
+      , "分け合うパワー！"
+      , "情熱のパワー！"
+      , "正義のパワー！"
+      , "（プリキュア・パーティアップ！）"
+      ]
   ]
  where
   preCureDeliciouStan'ByPartyGo = ["プリキュア！デリシャスタンバイ！", "パーティーゴー！"]
+
   nigiNigiKomeKomeHeartWoKomeKome = ["にぎにぎ！", komeKome, "ハートを！", komeKome]
   openPamPamSandPamPam = ["オープン！", pamPam, "サンド！", pamPam]
   kuruKuruMemMemMiracleMemMem = ["クルクル！", memMem, "ミラクル！", memMem]
+  fruitsFabulousOrder = ["フルーツ！", "ファビュラス・オーダー！"]
+  toppingBrilliantShineMore = ["トッピング！", "ブリリアント！", "シャインモア！"]
+
   sharin'Energy = "シェアリンエナジー！"
 
   komeKome = "（コメコメ！）"
@@ -131,7 +213,9 @@ transformations =
   wonton = "（ワンターン！）"
 
   girls1 = ["Yui", "Kokone", "Ran"]
+  girls2 = girls1 ++ ["Amane"]
   items1 = ["KomeKome", "PamPam", "MemMem"]
+  items2 = items1 ++ ["HeartFruitsPendant"]
 
 
 purifications :: [Purification]
@@ -162,6 +246,12 @@ purifications =
         ]
         ++ onakaIppaiGochisosamaDeshita
       )
+
+  --, mkPurification
+  --["CureFinale"]
+  --["CreamyFleuret"]
+  --[ "プリキュア！フィナーレブーケ！"
+  --]
 
   , mkPurification
       ["CurePrecious"]
@@ -200,6 +290,16 @@ purifications =
       )
 
   , mkPurification
+      ["CureFinale"]
+      ["CreamyFleuret"]
+      ( [ "クリーミーフルーレ！"
+        , "ブルーミン・ダンシンフルーツ！"
+        , "プリキュア！デリシャスフィナーレ・ファンファーレ！"
+        ]
+        ++ onakaIppaiGochisosamaDeshita
+      )
+
+  , mkPurification
       groupMembers1
       (replicate 3 "HeartJuicyMixer")
       ( [ "トリプルミックス！デリシャスチャージ！"
@@ -207,6 +307,15 @@ purifications =
         , "スパイシーフレイバー！"
         , "ヤムヤムフレイバー！"
         , "プリキュア！ミックスハート・アタック！"
+        ]
+        ++ onakaIppaiGochisosamaDeshita
+      )
+
+  , mkPurification
+      groupMembers_PartyUp
+      itemsLightMyDelicious
+      ( [ "心をひとつに！"
+        , "プリキュア！ライト・マイ・デリシャス！"
         ]
         ++ onakaIppaiGochisosamaDeshita
       )
@@ -227,8 +336,22 @@ purifications =
 groupMembers1 :: IsString s => [s]
 groupMembers1 = map fromString groupMembers1'
 
+groupMembers2 :: IsString s => [s]
+groupMembers2 = map fromString groupMembers2'
+
 groupMembers1' :: [String]
 groupMembers1' = ["CurePrecious", "CureSpicy", "CureYumYum"]
+
+groupMembers2' :: [String]
+groupMembers2' = groupMembers1' ++ ["CureFinale"]
+
+groupMembers_PartyUp :: IsString s => [s]
+groupMembers_PartyUp = map (fromString . (++ "_PartyUpStyle")) groupMembers2'
+
+
+itemsLightMyDelicious :: [IdAttachments]
+itemsLightMyDelicious = "KomeKome" : concat (replicate 4 ["PartyUpRing", "PartyCandleTact"])
+
 
 introducesHerselfAs_CurePrecious =
   "あつあつごはんで、みなぎるパワー！キュアプレシャス！おいしい笑顔で満たしてあげる！"
@@ -238,3 +361,6 @@ introducesHerselfAs_CureSpicy =
 
 introducesHerselfAs_CureYumYum =
   "きらめくヌードル・エモーション！キュアヤムヤム！おいしいの独り占め、ゆるさないよ！"
+
+introducesHerselfAs_CureFinale =
+  "ジェントルに、ゴージャスに！咲き誇るスウィートネス！キュアフィナーレ！食卓の最後を、このわたしが飾ろう。"